Move specific converters to utils
