+++ Sources/Common/Core/Injection/Swinject+Core.swift
@@ -10,17 +10,6 @@ extension Container {
     public func forward<T, U>(_ type: T.Type, to forwardType: U.Type) -> ServiceEntry<T> {
         return register(type) { $0.resolve(forwardType) as! T }
     }
-
-    public func registerListConverter<ItemConverter: ConverterProtocol>(
-        for itemConverterType: ItemConverter.Type
-    ) -> ServiceEntry<ListConverter<ItemConverter>> {
-        return autoregister(ListConverter<ItemConverter>.self, initializer: ListConverterImpl<ItemConverter>.init)
-    }
-
-    public func registerValueWrapperConverters<T: ValueWrapper>(for type: T.Type) where T.Value: Codable {
-        autoregister(Converter<WrapperDto<T.Value>, T>.self, initializer: ValueWrapperDomainConverterImpl<T>.init)
-        autoregister(Converter<T, WrapperDto<T.Value>>.self, initializer: ValueWrapperExternalConverterImpl<T>.init)
-    }
 }
 
 public extension Resolver {
